Restructured folders, fixed relative imports
