+++ python/cdhtools/plot_base.py
